Replace offset() by position()
The .position() method allows us to retrieve the current position of an element relative to the offset parent. Contrast this with .offset(), which retrieves the current position relative to the document. When positioning a new element near another one and within the same containing DOM element, .position() is the more useful. (cf. https://api.jquery.com/position/)

This prevents possible positioning issue if container manipulated by another script.